+++ android-app/app/src/main/java/com/swallaby/foodon/domain/calendar/model/RecommendFood.kt
@@ -5,6 +5,6 @@ import com.swallaby.foodon.domain.food.model.NutrientClaimInfo
 data class RecommendFood(
     val foodId: Long = 0,
     val name: String = "",
-    val kcal: Int = 0,
+    val kcal: Double = 0.0,
     val nutrientClaims: List<NutrientClaimInfo> = emptyList()
 )
